Convert from using webpack-dev-server to universal rendering
NOTE: Ran into issue where node was trying to build LESS files
* solved the issue by using https://github.com/morlay/babel-plugin-transform-require-ignore
