+++ front_end/.babelrc
@@ -1,11 +1,16 @@
 {
   "presets": [
-    "react",
-    ["es2015", {"modules": false, "loose": true}]
+    "react", ["es2015", {
+      "modules": false,
+      "loose": true
+    }]
   ],
   "env": {
     "test": {
       "plugins": ["transform-es2015-modules-commonjs"]
-    }
+    },
+    "server": {
+      "plugins": ["transform-es2015-modules-commonjs"]
+    },
   }
 }
